Move tesselation
