+++ distance3d/pressure_field/_contact_surface.py
@@ -5,24 +5,24 @@ from ..utils import transform_points, transform_directions
 class ContactSurface:
     def __init__(
             self, frame2world, intersection, contact_planes, contact_polygons,
-            contact_polygon_triangles, intersecting_tetrahedra1,
-            intersecting_tetrahedra2):
+            intersecting_tetrahedra1, intersecting_tetrahedra2):
         self.frame2world = frame2world
         self.intersection = intersection
         self.contact_planes = contact_planes
         self.contact_polygons = contact_polygons
-        self.contact_polygon_triangles = contact_polygon_triangles
         self.intersecting_tetrahedra1 = intersecting_tetrahedra1
         self.intersecting_tetrahedra2 = intersecting_tetrahedra2
 
         self.contact_areas = None
         self.contact_coms = None
         self.contact_forces = None
+        self.contact_polygon_triangles = None
 
-    def add_polygon_info(self, contact_areas, contact_coms, contact_forces):
+    def add_polygon_info(self, contact_areas, contact_coms, contact_forces, contact_polygon_triangles):
         self.contact_areas = contact_areas
         self.contact_coms = contact_coms
         self.contact_forces = contact_forces
+        self.contact_polygon_triangles = contact_polygon_triangles
 
     def _transform_to_world(self, frame2world):
         self.contact_polygons = [transform_points(frame2world, contact_polygon)
