+++ examples/visualizations/vis_tetrahedron_intersection.py
@@ -24,10 +24,9 @@ intersection, contact = pressure_field.intersect_tetrahedron_pair(
     tetrahedron1, epsilon1, X1, tetrahedron2, epsilon2, X2)
 assert intersection
 contact_plane_hnf, contact_polygon = contact
-triangles = pressure_field.tesselate_ordered_polygon(len(contact_polygon))
 
-intersection_com, force_vector, _ = pressure_field.compute_contact_force(
-    tetrahedron1, epsilon1, contact_plane_hnf, contact_polygon, triangles)
+intersection_com, force_vector, _, _ = pressure_field.compute_contact_force(
+    tetrahedron1, epsilon1, contact_plane_hnf, contact_polygon)
 
 fig = pv.figure()
 fig.scatter(tetrahedron1, s=0.01, c=(1, 0, 0))
